Fix vortex-tilt errors found during SHiELD testing
A lot of new code was added to diagnose vortex tilt using either
relative vorticity, wind circulation, gp height or temperature.
These changes were tested using both SHiELD and T-SHiELD NetCDF
data, and a number of fixes had to be made to the code.  In
addition, a change had to be made to the cwaitfor.c program to
fix an output formatting issue that only appeared once we
compiled with gnu instead of intel.

+++ code/src/tracker/gettrk_modules.f
@@ -311,7 +311,7 @@ c
         integer, parameter ::  vortex_max_levs=100 ! Max allowable
                                                 ! number of vertical
                                                 ! levs for vortex tilt
-        integer, save :: vortex_tilt_allow_thresh ! Max distance (km)
+        real, save :: vortex_tilt_allow_thresh  ! Max distance (km)
                                                 ! per mb of height 
                                                 ! difference allowed
                                                 ! between vortex fixes
@@ -341,7 +341,7 @@ c
           real, save, allocatable  ::  xtiltlon(:,:)
           real, save, allocatable  ::  xtiltlat(:,:)
           real, save, allocatable  ::  xtiltval(:,:)
-          real, save, allocatable  ::  xtilt_dist_flag(:,:)
+          integer, save, allocatable :: xtilt_dist_flag(:,:)
           integer, save, allocatable :: ifhours(:)  
           integer, save, allocatable :: iftotalmins(:)
           integer, save, allocatable :: ifclockmins(:)
